Move DurationConverters to javaapi
